requirements.txt
apache-airflow
pymongo
pathlib
jsonpath
pandas
numpy
gensim
nltk
elasticsearch>=6.0.0,<7.0.0